+++ Engine/Engine.Graphics/FlatBatch2D.cs
@@ -10,8 +10,8 @@ namespace Engine.Graphics
 			int count = LineVertices.Count;
 			LineVertices.Add(new VertexPositionColor(new Vector3(p1, depth), color));
 			LineVertices.Add(new VertexPositionColor(new Vector3(p2, depth), color));
-			LineIndices.Add((ushort)count);
-			LineIndices.Add((ushort)(count + 1));
+			LineIndices.Add(count);
+			LineIndices.Add(count + 1);
 		}
 
 		public void QueueLineStrip(IEnumerable<Vector2> points, float depth, Color color)
@@ -25,8 +25,8 @@ namespace Engine.Graphics
 			}
 			for (int i = 0; i < num - 1; i++)
 			{
-				LineIndices.Add((ushort)(count + i));
-				LineIndices.Add((ushort)(count + i + 1));
+				LineIndices.Add(count + i);
+				LineIndices.Add(count + i + 1);
 			}
 		}
 
@@ -37,14 +37,14 @@ namespace Engine.Graphics
 			LineVertices.Add(new VertexPositionColor(new Vector3(corner1.X, corner2.Y, depth), color));
 			LineVertices.Add(new VertexPositionColor(new Vector3(corner2.X, corner2.Y, depth), color));
 			LineVertices.Add(new VertexPositionColor(new Vector3(corner2.X, corner1.Y, depth), color));
-			LineIndices.Add((ushort)count);
-			LineIndices.Add((ushort)(count + 1));
-			LineIndices.Add((ushort)(count + 1));
-			LineIndices.Add((ushort)(count + 2));
-			LineIndices.Add((ushort)(count + 2));
-			LineIndices.Add((ushort)(count + 3));
-			LineIndices.Add((ushort)(count + 3));
-			LineIndices.Add((ushort)count);
+			LineIndices.Add(count);
+			LineIndices.Add(count + 1);
+			LineIndices.Add(count + 1);
+			LineIndices.Add(count + 2);
+			LineIndices.Add(count + 2);
+			LineIndices.Add(count + 3);
+			LineIndices.Add(count + 3);
+			LineIndices.Add(count);
 		}
 
 		public void QueueEllipse(Vector2 center, Vector2 radius, float depth, Color color, int sides = 32, float startAngle = 0f, float endAngle = (float)Math.PI * 2f)
@@ -103,9 +103,9 @@ namespace Engine.Graphics
 			TriangleVertices.Add(new VertexPositionColor(new Vector3(p1.X, p1.Y, depth), color));
 			TriangleVertices.Add(new VertexPositionColor(new Vector3(p2.X, p2.Y, depth), color));
 			TriangleVertices.Add(new VertexPositionColor(new Vector3(p3.X, p3.Y, depth), color));
-			TriangleIndices.Add((ushort)count);
-			TriangleIndices.Add((ushort)(count + 1));
-			TriangleIndices.Add((ushort)(count + 2));
+			TriangleIndices.Add(count);
+			TriangleIndices.Add(count + 1);
+			TriangleIndices.Add(count + 2);
 		}
 
 		public void QueueTriangle(Vector2 p1, Vector2 p2, Vector2 p3, float depth, Color color1, Color color2, Color color3)
@@ -114,9 +114,9 @@ namespace Engine.Graphics
 			TriangleVertices.Add(new VertexPositionColor(new Vector3(p1.X, p1.Y, depth), color1));
 			TriangleVertices.Add(new VertexPositionColor(new Vector3(p2.X, p2.Y, depth), color2));
 			TriangleVertices.Add(new VertexPositionColor(new Vector3(p3.X, p3.Y, depth), color3));
-			TriangleIndices.Add((ushort)count);
-			TriangleIndices.Add((ushort)(count + 1));
-			TriangleIndices.Add((ushort)(count + 2));
+			TriangleIndices.Add(count);
+			TriangleIndices.Add(count + 1);
+			TriangleIndices.Add(count + 2);
 		}
 
 		public void QueueQuad(Vector2 corner1, Vector2 corner2, float depth, Color color)
@@ -126,12 +126,12 @@ namespace Engine.Graphics
 			TriangleVertices.Add(new VertexPositionColor(new Vector3(corner1.X, corner2.Y, depth), color));
 			TriangleVertices.Add(new VertexPositionColor(new Vector3(corner2.X, corner2.Y, depth), color));
 			TriangleVertices.Add(new VertexPositionColor(new Vector3(corner2.X, corner1.Y, depth), color));
-			TriangleIndices.Add((ushort)count);
-			TriangleIndices.Add((ushort)(count + 1));
-			TriangleIndices.Add((ushort)(count + 2));
-			TriangleIndices.Add((ushort)(count + 2));
-			TriangleIndices.Add((ushort)(count + 3));
-			TriangleIndices.Add((ushort)count);
+			TriangleIndices.Add(count);
+			TriangleIndices.Add(count + 1);
+			TriangleIndices.Add(count + 2);
+			TriangleIndices.Add(count + 2);
+			TriangleIndices.Add(count + 3);
+			TriangleIndices.Add(count);
 		}
 
 		public void QueueQuad(Vector2 p1, Vector2 p2, Vector2 p3, Vector2 p4, float depth, Color color)
@@ -144,12 +144,12 @@ namespace Engine.Graphics
 			TriangleVertices.Array[count + 3] = new VertexPositionColor(new Vector3(p4.X, p4.Y, depth), color);
 			int count2 = TriangleIndices.Count;
 			TriangleIndices.Count += 6;
-			TriangleIndices.Array[count2] = (ushort)count;
-			TriangleIndices.Array[count2 + 1] = (ushort)(count + 1);
-			TriangleIndices.Array[count2 + 2] = (ushort)(count + 2);
-			TriangleIndices.Array[count2 + 3] = (ushort)(count + 2);
-			TriangleIndices.Array[count2 + 4] = (ushort)(count + 3);
-			TriangleIndices.Array[count2 + 5] = (ushort)count;
+			TriangleIndices.Array[count2] = count;
+			TriangleIndices.Array[count2 + 1] = count + 1;
+			TriangleIndices.Array[count2 + 2] = count + 2;
+			TriangleIndices.Array[count2 + 3] = count + 2;
+			TriangleIndices.Array[count2 + 4] = count + 3;
+			TriangleIndices.Array[count2 + 5] = count;
 		}
 
 		public void QueueQuad(Vector2 p1, Vector2 p2, Vector2 p3, Vector2 p4, float depth, Color color1, Color color2, Color color3, Color color4)
@@ -162,12 +162,12 @@ namespace Engine.Graphics
 			TriangleVertices.Array[count + 3] = new VertexPositionColor(new Vector3(p4.X, p4.Y, depth), color4);
 			int count2 = TriangleIndices.Count;
 			TriangleIndices.Count += 6;
-			TriangleIndices.Array[count2] = (ushort)count;
-			TriangleIndices.Array[count2 + 1] = (ushort)(count + 1);
-			TriangleIndices.Array[count2 + 2] = (ushort)(count + 2);
-			TriangleIndices.Array[count2 + 3] = (ushort)(count + 2);
-			TriangleIndices.Array[count2 + 4] = (ushort)(count + 3);
-			TriangleIndices.Array[count2 + 5] = (ushort)count;
+			TriangleIndices.Array[count2] = count;
+			TriangleIndices.Array[count2 + 1] = count + 1;
+			TriangleIndices.Array[count2 + 2] = count + 2;
+			TriangleIndices.Array[count2 + 3] = count + 2;
+			TriangleIndices.Array[count2 + 4] = count + 3;
+			TriangleIndices.Array[count2 + 5] = count;
 		}
 
 		public void TransformLines(Matrix matrix, int start = 0, int end = -1)
